transaction_verification/requirements.txt
grpcio
grpcio-tools
protobuf==5.29.0
watchdog==6.0.0
